Refactor messages into separate component
